requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
pandas>=2.0.0
pandas-ta>=0.3.14b
scipy>=1.11.0
numpy>=1.24.0
yfinance>=0.2.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
